Added the placement indicator
A grey tile (placeholder) will appear in the location where the new tile will be placed. If no tiles are there, it will appear on the bottom layer of the grid. Otherwise, if the mouse is ontop of a tile, the grey tile will appear above. If the mouse is on the side of a tile, the grey tile will appear one unit in the normal direction of the side of that tile.